Add files via upload
This is a simple demo of how HttpServlets work in Java. The example shows an index page with signup and login options. Once you click sign up, a form shows where you entered your data. If every entry is correct and you hit register you will be signed up and directed to the login page through the server along with a successful registration message. After login in you will be redirected to a page with a welcome message and logout option.